requirements.txt
pyrogram
pymongo
pymongo[srv]
dnspython
motor
chardet
python-dotenv
tgcrypto
